backend/requirements.txt
# Patella Care AI - Backend
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
pydantic>=2.0.0
ultralytics>=8.0.0
